exported formState reducer into it's own js file
